% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_anova.R
\name{auto_anova}
\alias{auto_anova}
\title{Auto-select ANOVA path and plot}
\usage{
auto_anova(
  formula,
  data,
  which_factor = NULL,
  ci_level = 0.95,
  point_alpha = 0.5,
  jitter_width = 0.12,
  show_means = c("point", "point+ci", "none"),
  theme_base = ggplot2::theme_bw(),
  normality_method = c("auto", "subsample", "ad", "jb", "none", "shapiro"),
  shapiro_limit = 5000,
  seed = 42,
  pairwise_adjust = c("sidak", "tukey", "bonferroni", "holm", "BH"),
  max_groups_for_letters = 25,
  alpha_letters = 0.05,
  prefer_welch_when_large_n = TRUE,
  min_n_for_welch = 15,
  subtitle_style = c("full", "compact", "minimal")
)
}
\description{
Auto-select ANOVA path and plot
}
