% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_interval.R
\name{confidence_interval}
\alias{confidence_interval}
\title{Calculating the confidence interval when mean, sd, n, and CI are known}
\usage{
confidence_interval(percent = 0.95, s, n = 1, data = NULL)
}
\arguments{
\item{percent}{Numeric; A number between 0 and 1 corresponding to the confidence interval level. If not defined, defaults to 0.95.}

\item{s}{Numeric; the standard deviation for the given data.}

\item{n}{Numeric; the number of records in a dataset; if blank, defaults to 1.}

\item{data}{Numeric vector; a vector of numberic data; if used, arguments for s and n ignored.}
}
\value{
A value for the confidence interval, given the value on either side of the mean that encompasses the interval.
}
\description{
Calculating the confidence interval when mean, sd, n, and CI are known
}
