% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_test.R
\name{correlation_test}
\alias{correlation_test}
\title{Calculating the coefficient of variation for a numeric vector.}
\usage{
correlation_test(x = NA, y = NA, method = "pearson")
}
\arguments{
\item{x}{Numeric; numeric vector.}

\item{y}{Numeric; numeric vector.}

\item{method}{Character; method of test to perform. Options are "pearson", "kendall", "spearman". Defaults to "pearson".}
}
\value{
A printout of values needed for correlations, including values for testing them and p values for reporting.
}
\description{
Calculating the coefficient of variation for a numeric vector.
}
