% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy_plotter.R
\name{easy_plotter}
\alias{easy_plotter}
\title{Easy Plotting with Error Bars}
\usage{
easy_plotter(
  data = NULL,
  response = NULL,
  group = NULL,
  error_measure = "se",
  theme_base = theme_classic()
)
}
\arguments{
\item{data}{A data frame.}

\item{response}{The response variable.}

\item{group}{Character; grouping factor.}

\item{error_measure}{The measurement used for the error bars. Options include "se" for standard error, "sd" for standard deviation, and "ci" for 95\% confidence interval.}

\item{theme_base}{ggplot2 theme.}
}
\value{
A plot with the desired parameters.
}
\description{
Creates simple ggplots with data with error bars of your choice shown.
}
