% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_mode.R
\name{find_mode}
\alias{find_mode}
\title{Finding the mode for a numeric vector.
Based on Statology function
define function to calculate mode
works on vectors of data}
\usage{
find_mode(x)
}
\arguments{
\item{x}{Numeric; numeric vector.}
}
\value{
A value for the mode of a data string. Indicates if no mode is found.
}
\description{
Finding the mode for a numeric vector.
Based on Statology function
define function to calculate mode
works on vectors of data
}
