% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_p.R
\name{p_rounder}
\alias{p_rounder}
\title{Calculating the coefficient of variation for a numeric vector.}
\usage{
p_rounder(p)
}
\arguments{
\item{p}{Numeric; numeric vector.}
}
\value{
Correct rounding for p values based on Ornithological Applications.
}
\description{
Calculating the coefficient of variation for a numeric vector.
}
