% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_oneway.R
\name{run_factorial}
\alias{run_factorial}
\title{Factorial ANOVA (student-friendly)}
\usage{
run_factorial(
  data,
  dv,
  between,
  which_factor = NULL,
  adjust = c("sidak", "tukey", "holm", "bonferroni", "BH"),
  show_means = c("point", "point+ci", "none"),
  theme_base = ggplot2::theme_bw()
)
}
\description{
Runs factorial ANOVA or ART as needed and returns tidy tables + plot.
}
