% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se}
\alias{se}
\title{Calculating standard error from a string of numbers}
\usage{
se(x)
}
\arguments{
\item{x}{Numeric; numeric vector.}
}
\value{
A value for the standard error.
}
\description{
Calculating standard error from a string of numbers
}
