% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore.R
\name{zscore}
\alias{zscore}
\title{Calculating the Z score when the sample mean, population mean, and standard deviation are known.}
\usage{
zscore(xbar, mu, sd.x, n = 1)
}
\arguments{
\item{xbar}{Numeric; value representing the sample mean.}

\item{mu}{Numeric; value representing the population mean.}

\item{sd.x}{Numeric; value representing the standard deviation of the sample, if known. Can be replaced with population mean in some scenarios.}

\item{n}{Numeric; value representing the sample size. Defaults to 1 if not defined.}
}
\value{
A value for the Z score.
}
\description{
Calculating the Z score when the sample mean, population mean, and standard deviation are known.
}
