% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R, R/run_oneway.R
\name{cv}
\alias{cv}
\title{Calculating the coefficient of variation for a numeric vector.}
\usage{
cv(x)

cv(x)
}
\arguments{
\item{x}{Numeric; numeric vector.}
}
\value{
A value for the coefficient of variation for a given numeric vector.

A value for the coefficient of variation for a given numeric vector.
}
\description{
Calculating the coefficient of variation for a numeric vector.

Calculating the coefficient of variation for a numeric vector.
}
