% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_oneway.R
\name{run_oneway}
\alias{run_oneway}
\title{One-way ANOVA (student-friendly)}
\usage{
run_oneway(
  data,
  dv,
  group,
  adjust = c("tukey", "sidak", "holm", "bonferroni", "BH"),
  show_means = c("point", "point+ci", "none"),
  theme_base = ggplot2::theme_bw()
)
}
\description{
Runs one-way ANOVA, Welch, or Kruskal–Wallis automatically and returns
test info, tidy tables, and a plot.
}
